Use target_compile_features to request C++11
Currently, the samples don't build because I forgot to add the -gnu++11
flag to the CMakeLists.txt. Rather than do that (and add logic to detect
MSVC), I'd prefer to upgrade to CMake 3.1 and use
target_compile_features().

target_compile_features() provides a superior way to control which
language standard is needed, because it propagates on public linking, and
figures out which flag your compiler needs for you.

Specifically, if target `kj` requires a C++11 feature publicly, then all
targets which link to `kj` will also require that C++11 feature, and get
the -std=gnu++XX flag for free. If that target is a library, such as
`kj-async`, and `kj-async` links publicly to `kj`, the requirement is also
transitive to all targets which link to `kj-async`.

Note that CMake's default behavior is to request compiler-specific
extensions, such as those provided by the -std=gnu++11 flag. You must
specifically opt out of these extensions. I'm not aware of any way to
propagate this nuance in CMake, but added a check to at least prevent
users from compiling Cap'n Proto without extensions.

With this change, the c++/samples/CMakeLists.txt will build with MinGW
with no change beyond requiring a later version of CMake.